Now Drag_Mode should be supported with SAME_LEVEL
